Validate events; make event numbers longs

+++ platform-sdk/swirlds-common/src/test/java/com/swirlds/common/wiring/components/Event.java
@@ -1,12 +1,12 @@
 package com.swirlds.common.wiring.components;
 
 public final class Event {
-    private int number = -1; // We'll let the orphan buffer assign this, although I think consensus actually does
+    private long number = -1; // We'll let the orphan buffer assign this, although I think consensus actually does
     private final byte[] data = new byte[1024 * 32]; // Just gotta have some bytes. Whatever.
 
     public Event() { }
 
-    void reset(int number) {
+    void reset(long number) {
         this.number = number;
     }
 
@@ -15,7 +15,7 @@ public final class Event {
         return "Event {number=" + number + "}";
     }
 
-    public int number() {
+    public long number() {
         return number;
     }
 }
